Setup basic virtual file system in server
Not sure if this is the best way to do this. Although a virtual file
system could be useful during testing, I could use node:fs to deal with
actual files during production. I'll also need to sync the directory
tree structure at a later point with liveblocks to showcase the live
editing and live avatars.